Script restructure, Collider improvements
Complete restructure of script. Divided into multiple manageable scripts. Player collider improved, allows for some error.
