+++ index.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="game-styles.css">
     <link href="https://fonts.googleapis.com/css2?family=Press+Start+2P&display=swap" rel="stylesheet">
     <script src="https://cdn.jsdelivr.net/npm/phaser@3.60.0/dist/phaser.min.js"></script>
-    <script defer src="baseball-running-game.js"></script>
+    <script type="module" src="baseball-running-game.js"></script>
 </head>
 <body>
     <div id="game-container"></div>
